SLIDER-228 Lift maven-compiler-plugin definition to pluginManagement and add enforcer plugin to verify minimum jdk
We set the groovy-eclipse-compiler on the maven-compiler-plugin to get compilation of both java
and groovy files. It doesn't appear that the groovy-eclipse-compiler will fail when the target
JDK version is different than the actual version of java on the system (like the default compiler
used by the maven-compiler-plugin does).

Sets up an enforcer plugin rule that asserts the JDK minimum version instead of relying
on the compiler plugin to appropriately warn the user when <1.7 is used.