SLIDER-1039 Run python tests from target directory
We shouldn't be running any python tests from the src
directory as this is for the clean copy of the code. This
should be done in the target directory. Uses the maven-resources-plugin
to copy the src and test python code into target and changes the
configuration on the exec plugin to specify the right paths
rooted in target.

Also removes some duplicative configuration when building on *nixes.

This does not address the glaring problem of slider-assembly
reaching into slider-agent directly to access python files.